requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
Flask==2.2.2
Flask-WTF==1.0.1
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mysql-connector-python==8.0.31
names==0.3.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
patchworklib==0.4.7
Pillow==9.3.0
plotly==5.11.0
protobuf==3.20.1
pyarrow==10.0.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
WTForms==3.0.1
xlrd==2.0.1
zipp==3.10.0
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
Flask==2.2.2
Flask-WTF==1.0.1
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mysql-connector-python==8.0.31
names==0.3.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
patchworklib==0.4.7
Pillow==9.3.0
plotly==5.11.0
protobuf==3.20.1
pyarrow==10.0.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
WTForms==3.0.1
xlrd==2.0.1
zipp==3.10.0
